requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
multi-key-dict==2.0.3
pbr==3.1.0
python-jenkins==0.4.14
six==1.10.0
Werkzeug==0.12.2
